requirements.txt
torch>=2.1.0
torchvision>=0.16.0
datasets>=2.18.0
pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
tqdm>=4.66.0
opencv-python>=4.8.0
grad-cam>=1.5.0
PyYAML>=6.0.0
